+++ src/Sentry.Unity.Editor/ConfigurationWindow/DebugSymbolsTab.cs
@@ -19,10 +19,6 @@ namespace Sentry.Unity.Editor.ConfigurationWindow
 
             EditorGUILayout.EndToggleGroup();
 
-            cliOptions.Url = EditorGUILayout.TextField(
-                new GUIContent("Sentry URL", "Fully qualified URL to the Sentry server (defaults to https://sentry.io/)"),
-                cliOptions.Url);
-
             cliOptions.Auth = EditorGUILayout.TextField(
                 new GUIContent("Auth Token", "The authorization token from your user settings in Sentry"),
                 cliOptions.Auth);
@@ -34,6 +30,11 @@ namespace Sentry.Unity.Editor.ConfigurationWindow
             cliOptions.Project = EditorGUILayout.TextField(
                 new GUIContent("Project Name", "The project name in Sentry"),
                 cliOptions.Project);
+
+            cliOptions.UrlOverride = EditorGUILayout.TextField(
+                new GUIContent("Override Sentry URL", "Fully qualified URL to the Sentry server " +
+                                                      "(defaults to https://sentry.io/)"),
+                cliOptions.UrlOverride);
         }
     }
 }